Initial GitHub Commit

+++ oscillatorBase.pde
@@ -1,15 +1,15 @@
 class oscillatorBase{
   float frequency;
-  float phase;
+  long phase;
   oscillatorBase(float _frequency,float _phase){
     frequency = _frequency;
-    phase = _phase;
+    phase = (long)_phase;
   }
   void setFreq(float _frequency){
     frequency = _frequency;
   }
   void setPhase(float _phase){
-    phase = _phase;
+    phase = (long)_phase;
   }
   float getFreq(){
     return frequency;
